Update for SilverStripe 5.4 with Pickr color picker
- Migrated from SS3 to SS5.4 with proper namespacing (Colymba\ColorField\)
- Replaced jQuery MiniColors with Pickr (vanilla JS, no dependencies)
- Implemented bi-directional HEX/RGB/Alpha editing
- Added color preview with transparency grid
- Maintained RRGGBBAAA database format compatibility (6 hex + 3 alpha digits)
- Updated folder structure from /code/ to /src/ (PSR-4)
- Added Pickr color picker with inline mode and color swatches
- Full SilverStripe 5.4+ CMS integration tested and working
