+++ _config.php
@@ -1,7 +1,3 @@
 <?php
-//define global path to Components' root folder
-if(!defined('SS_COLORFIELD'))
-{
-  $folder = rtrim(basename(dirname(__FILE__)));
-  define('SS_COLORFIELD', $folder);
-}
+
+// SilverStripe 5.x configuration is handled via YAML in _config/config.yml
